Adjust LoginActivity Layout
This commit modifies the layout XML file for the LoginActivity to improve visual consistency and alignment. Changes include adjusting the dimensions of the welcome message TextView and the profile image, as well as refining the margins for better spacing. Additionally, the maxLength attribute for the email TextInputEditText is removed to allow for flexibility in email length input.