Add FloatingActionButton to FitnessActivity Layout
This commit adds a FloatingActionButton to the layout XML file for the FitnessActivity. The FloatingActionButton serves as an intuitive UI element for users to add new fitness activities. It is positioned at the bottom end of the screen for easy access and is styled with a green color tint. The FloatingActionButton is represented by an icon depicting a plus sign.