Adjust LoginActivity Layout
This commit adjusts the layout XML file for the LoginActivity to improve visual consistency and alignment. Changes include refining the dimensions and margins of UI elements such as TextViews, ImageViews, and TextInputLayouts. Additionally, the maxLength attribute for the email TextInputEditText is removed to allow for flexibility in email length input.