Implement FloatingActionButton functionality in HomeFragment
This commit adds functionality to the HomeFragment by implementing a FloatingActionButton (FAB) that allows users to navigate to the AddFitnessActivity. Upon inflation of the fragment layout, the FloatingActionButton is located and assigned a click listener. When clicked, it triggers the navigation to the AddFitnessActivity using an Intent. This enhancement improves the user experience by providing a convenient way to add fitness activities directly from the home screen.