Add vector drawable for profile icon
This commit introduces a vector drawable file named "profile_icon.xml" containing the path data for a profile icon. The icon is designed to have a height and width of 24dp and is tinted with the color #2196F3. This drawable can be used within the app's user interface to represent profile-related actions or elements, enhancing the visual appeal and consistency of the app's design.