Introduce vector drawable for navigation arrow icon
This commit adds a vector drawable file named "navigation_arrow_icon.xml", which depicts a navigation arrow. The drawable has a height and width of 24dp and is tinted with the color #2196F3. It consists of a path representing the shape of a directional arrow. This drawable can be utilized within the app's user interface to indicate navigation-related actions or elements, contributing to a cohesive and intuitive user experience.